STM32Cube_FW_F4_V1.4.0 Released 2014-12-26
STM32CubeF4 Firmware Package V1.4.0 / December 26, 2014

Main Changes
* Maintenance release to fix known defects and several
  enhancements implementation
* HAL Macros and literals renaming to ensure compatibles across
  STM32 series, backward compatibility maintained thanks to new
  added file stm32_hal_legacy.h under /Inc/Legacy
* Use latest MW version: USB Host V3.2.0, USB Device V2.3.0,
  STemWin V5.26, FatFs R0.10b and FreeRTOS V8.1.2
* HAL and USB Libraries update to be C++ compliant
* BSP update to fix compilation issues under Linux and Mac OS

+++ Projects/STM324x9I_EVAL/Demonstrations/MB1063/Core/Src/k_sd_diskio.c
@@ -2,8 +2,8 @@
   ******************************************************************************
   * @file    k_sd_diskio.c
   * @author  MCD Application Team
-  * @version V1.1.0
-  * @date    26-June-2014   
+  * @version V1.2.0
+  * @date    26-December-2014   
   * @brief   kernel SD Disk I/O driver
   ******************************************************************************
   * @attention
@@ -39,14 +39,14 @@
 static volatile DSTATUS Stat = STA_NOINIT;
 
 /* Private function prototypes -----------------------------------------------*/
-DSTATUS SD_initialize (void);
-DSTATUS SD_status (void);
-DRESULT SD_read (BYTE*, DWORD, BYTE);
+DSTATUS SD_initialize(void);
+DSTATUS SD_status(void);
+DRESULT SD_read(BYTE*, DWORD, UINT);
 #if _USE_WRITE == 1
-  DRESULT SD_write (const BYTE*, DWORD, BYTE);
+  DRESULT SD_write(const BYTE*, DWORD, UINT);
 #endif /* _USE_WRITE == 1 */
 #if _USE_IOCTL == 1
-  DRESULT SD_ioctl (BYTE, void*);
+  DRESULT SD_ioctl(BYTE, void*);
 #endif  /* _USE_IOCTL == 1 */
   
 Diskio_drvTypeDef  SD_Driver =
@@ -107,7 +107,7 @@ DSTATUS SD_status(void)
   * @param  count: Number of sectors to read (1..128)
   * @retval DRESULT: Operation result
   */
-DRESULT SD_read(BYTE *buff, DWORD sector, BYTE count)
+DRESULT SD_read(BYTE *buff, DWORD sector, UINT count)
 {
   DRESULT res = RES_OK;
   __disable_irq();
@@ -131,7 +131,7 @@ DRESULT SD_read(BYTE *buff, DWORD sector, BYTE count)
   * @retval DRESULT: Operation result
   */
 #if _USE_WRITE == 1
-DRESULT SD_write(const BYTE *buff, DWORD sector, BYTE count)
+DRESULT SD_write(const BYTE *buff, DWORD sector, UINT count)
 {
   DRESULT res = RES_OK;
   __disable_irq();
@@ -197,4 +197,3 @@ DRESULT SD_ioctl(BYTE cmd, void *buff)
 #endif /* _USE_IOCTL == 1 */
   
 /************************ (C) COPYRIGHT STMicroelectronics *****END OF FILE****/
-
